+++ b4n-tui/tui.rs
@@ -1,5 +1,5 @@
 use anyhow::Result;
-use b4n_config::keys::KeyCombination;
+use b4n_config::keys::{KeyCombination, KeyCommand};
 use crossterm::cursor::SetCursorStyle;
 use crossterm::event::{DisableMouseCapture, EnableMouseCapture, KeyModifiers, MouseButton};
 use futures::{FutureExt, StreamExt};
@@ -76,6 +76,7 @@ pub enum MouseEventKind {
 pub enum TuiEvent {
     Key(KeyCombination),
     Mouse(MouseEvent),
+    Command(KeyCommand),
 }
 
 impl TuiEvent {
@@ -116,7 +117,7 @@ impl TuiEvent {
     pub fn position(&self) -> Option<Position> {
         match self {
             TuiEvent::Mouse(mouse) => Some(Position::new(mouse.column, mouse.row)),
-            TuiEvent::Key(_) => None,
+            TuiEvent::Key(_) | TuiEvent::Command(_) => None,
         }
     }
 }